requirements.txt
Flask
Flask-SQLAlchemy
Flask-Bcrypt
pyserial